+++ variables.tf
@@ -372,7 +372,7 @@ variable "ami_id" {
 variable "graphdb_version" {
   description = "GraphDB version"
   type        = string
-  default     = "11.1.3"
+  default     = "11.2.0"
   nullable    = false
 }